+++ libs/opengl/include/mrpt/opengl/CTexturedPlane.h
@@ -10,6 +10,7 @@
 
 #include <mrpt/math/TPolygonWithPlane.h>
 #include <mrpt/opengl/CRenderizableShaderTexturedTriangles.h>
+#include <mrpt/opengl/CRenderizableShaderTriangles.h>
 
 namespace mrpt::opengl
 {
@@ -18,7 +19,8 @@ namespace mrpt::opengl
  *  \sa opengl::COpenGLScene
  * \ingroup mrpt_opengl_grp
  */
-class CTexturedPlane : public CRenderizableShaderTexturedTriangles
+class CTexturedPlane : public CRenderizableShaderTexturedTriangles,
+					   public CRenderizableShaderTriangles
 {
 	DEFINE_SERIALIZABLE(CTexturedPlane, mrpt::opengl)
 
@@ -34,7 +36,20 @@ class CTexturedPlane : public CRenderizableShaderTexturedTriangles
    public:
 	/** @name Renderizable shader API virtual methods
 	 * @{ */
+	void render(const RenderContext& rc) const override;
+	void renderUpdateBuffers() const override;
 	virtual void onUpdateBuffers_TexturedTriangles() override;
+	virtual void onUpdateBuffers_Triangles() override;
+	virtual shader_list_t requiredShaders() const override
+	{
+		return {
+			DefaultShaderID::TRIANGLES, DefaultShaderID::TEXTURED_TRIANGLES};
+	}
+	void freeOpenGLResources() override
+	{
+		CRenderizableShaderTriangles::freeOpenGLResources();
+		CRenderizableShaderTexturedTriangles::freeOpenGLResources();
+	}
 	/** @} */
 
 	CTexturedPlane(