+++ doc/source/doxygen-docs/changelog.md
@@ -52,6 +52,7 @@
       - mrpt::opengl::CRenderizableShaderTexturedTriangles::cullFaces()
     - Remove specular light effects in the default shaders, to fix buggy behavior.
     - **[API change]** New mrpt::opengl::Visualizable interface replaces former getAs3DObject() in all mrpt::maps and mrpt::poses classes with an uniform API, avoiding shared_ptr if possible.
+    - mrpt::opengl::CTexturedPlane now more efficiently renders as plain triangles if no texture has been assigned.
   - \ref mrpt_poses_grp
     - New function mrpt::poses::sensor_poses_from_yaml()
     - New header `<mrpt/poses/gtsam_wrappers.h>`, see \ref mrpt_gtsam_wrappers